+++ examples/panelhardwarescroll/panelhardwarescroll.ino
@@ -9,7 +9,7 @@
 CRGB leds[NUMSTRIPS * NUM_LEDS_PER_STRIP];
 
 
-int pins[16] = {0, 2, 4, 5, 12, 13, 14, 15, 16, 18, 19, 21, 22, 23, 25, 26};
+uint8_t pins[16] = {0, 2, 4, 5, 12, 13, 14, 15, 16, 18, 19, 21, 22, 23, 25, 26};
 
 OffsetDisplay offd;
 I2SClocklessLedDriver driver;
